follow

+++ main.py
@@ -73,7 +73,6 @@ def getSessionId(request : Request, session_key : str) :
     else:
         return {"message": "Session value not found"}
 
-
 @app.get("/api/login")
 def login(response: Response, id : str, pw : str) :
     result1 = sql.select(f"SELECT count(*) as count FROM user WHERE id='{id}' AND pw ='{pw}'")
@@ -96,12 +95,12 @@ def login(id : str, pw : str) :
 
 @app.get("/api/article")
 def getAllArticle() :
-    result = sql.select(f"SELECT user_id, content, time FROM article ORDER BY time DESC")
+    result = sql.select(f"SELECT * FROM article ORDER BY time DESC")
     return result
 
 @app.get("/api/article/{user_id}")
 def getUserArticle(user_id: int) :
-    result = sql.select(f"SELECT user_id, content, time FROM article WHERE user_id = {user_id} ORDER BY time DESC")
+    result = sql.select(f"SELECT * FROM article WHERE user_id = {user_id} ORDER BY time DESC")
     return result
 
 class InsertArticle(BaseModel):
@@ -125,18 +124,28 @@ class UpdateArticle(BaseModel):
 def updateArticle(request : Request, parameter: UpdateArticle, article_id : int) :
     session_id = getSessionId(request, "session")
 
-    sql.update(f"UPDATE article SET content = '{parameter.content}' WHERE idx = {article_id} AND user_id = '{session_id}'")
-    return {"result" : "success"}
+    result = sql.select(f"SELECT COUNT(*) as count FROM article WHERE idx = {article_id} AND user_id = '{session_id}'")
+
+    if result[0]['count'] != 0 :
+        sql.update(f"UPDATE article SET content = '{parameter.content}' WHERE idx = {article_id} AND user_id = '{session_id}'")
+        return {"result" : "success"}
+    else : 
+        return {"result" : "fail"}
 
 @app.delete("/api/article/{article_id}")
-def insertArticle(request : Request, article_id : int) :
+def deleteArticle(request : Request, article_id : int) :
     session_id = getSessionId(request, "session")
 
-    sql.delete(f"DELETE FROM article WHERE idx = '{article_id}' AND user_id='{session_id}'")
-    return {"result" : "success"}
+    result = sql.select(f"SELECT COUNT(*) as count FROM article WHERE idx = {article_id} AND user_id = '{session_id}'")
+
+    if result[0]['count'] != 0 :
+        sql.delete(f"DELETE FROM article WHERE idx = '{article_id}' AND user_id='{session_id}'")
+        return {"result" : "success"}
+    else : 
+        return {"result" : "fail"}
 
 @app.get("/api/article/like/{article_id}")
-def getArticleLike(article_id : int) :
+def getArticleLikeCount(article_id : int) :
     result = sql.select(f"SELECT COUNT(*) as count FROM article_like WHERE article_id={article_id}")
     
     return {"result" : result[0]['count']}
@@ -148,19 +157,19 @@ def articleLike(request : Request, article_id : int) :
     result = sql.select(f"SELECT COUNT(*) as count FROM article_like WHERE article_id={article_id} AND like_id='{session_id}'")
     
     if result[0]['count'] == 0 :
-        sql.insert(f"INSERT INTO article_like(like_id, article_id) VALUES ('{session_id}', {article_id})")
+        sql.insert(f"INSERT INTO article_like(article_id, like_id) VALUES ({article_id}, '{session_id}')")
         return {"result" : "like"}
     else : 
         sql.delete(f"DELETE FROM article_like WHERE article_id={article_id} AND like_id='{session_id}'")
         return {"result" : "dislike"}
 
-class InsertComment(BaseModel):
-    content: str
-
 @app.get("/api/comment/{article_id}")
 def getUserComment(article_id: int) :
-    result = sql.select(f"SELECT user_id, content, time FROM comment WHERE article_id = {article_id} ORDER BY time DESC")
+    result = sql.select(f"SELECT * FROM comment WHERE article_id = {article_id} ORDER BY time DESC")
     return result
+
+class InsertComment(BaseModel):
+    content: str
 
 @app.post("/api/comment/{article_id}")
 def insertComment(request : Request, parameter: InsertComment, article_id : int) :
@@ -173,19 +182,61 @@ def insertComment(request : Request, parameter: InsertComment, article_id : int)
     except : 
         return {"result" : "fail"}
 
-class UpdateArticle(BaseModel):
+class UpdateComment(BaseModel):
     content: str
 
-# @app.put("/api/comment/{article_id}")
-# def updateComment(request : Request, parameter: UpdateArticle, article_id : int) :
-#     session_id = getSessionId(request, "session")
+@app.put("/api/comment/{comment_id}")
+def updateComment(request : Request, parameter: UpdateComment, comment_id : int) :
+    session_id = getSessionId(request, "session")
 
-#     sql.update(f"UPDATE article SET content = '{parameter.content}' WHERE idx = {article_id} AND user_id = '{session_id}'")
-#     return {"result" : "success"}
+    result = sql.select(f"SELECT COUNT(*) as count FROM comment WHERE idx = {comment_id} AND user_id = '{session_id}'")
 
-# @app.delete("/api/comment/{article_id}")
-# def insertArticle(request : Request, article_id : int) :
-#     session_id = getSessionId(request, "session")
+    if result[0]['count'] != 0 :
+        sql.update(f"UPDATE comment SET content = '{parameter.content}' WHERE idx = {comment_id} AND user_id = '{session_id}'")
+        return {"result" : "success"}
+    else : 
+        return {"result" : "fail"}
 
-#     sql.delete(f"DELETE FROM article WHERE idx = '{article_id}' AND user_id='{session_id}'")
-#     return {"result" : "success"}
+@app.delete("/api/comment/{comment_id}")
+def insertArticle(request : Request, comment_id : int) :
+    session_id = getSessionId(request, "session")
+
+    result = sql.select(f"SELECT COUNT(*) as count FROM comment WHERE idx = {comment_id} AND user_id = '{session_id}'")
+
+    if result[0]['count'] != 0 :
+        sql.delete(f"DELETE FROM comment WHERE idx = '{comment_id}' AND user_id='{session_id}'")
+        return {"result" : "success"}
+    else : 
+        return {"result" : "fail"}
+
+@app.get("/api/comment/like/{comment_id}")
+def getCommentLikeCount(comment_id : int) :
+    result = sql.select(f"SELECT COUNT(*) as count FROM comment_like WHERE comment_id={comment_id}")
+    
+    return {"result" : result[0]['count']}
+
+@app.post("/api/comment/like/{comment_id}")
+def articleLike(request : Request, comment_id : int) :
+    session_id = getSessionId(request, "session")
+
+    result = sql.select(f"SELECT COUNT(*) as count FROM comment_like WHERE comment_id={comment_id} AND like_id='{session_id}'")
+    
+    if result[0]['count'] == 0 :
+        sql.insert(f"INSERT INTO comment_like(comment_id, like_id) VALUES ({comment_id}, '{session_id}')")
+        return {"result" : "like"}
+    else : 
+        sql.delete(f"DELETE FROM comment_like WHERE comment_id={comment_id} AND like_id='{session_id}'")
+        return {"result" : "dislike"}
+
+@app.post("/api/follow/{following_id}")
+def insertArticle(request : Request, following_id : str) :
+    session_id = getSessionId(request, "session")
+
+    result = sql.select(f"SELECT COUNT(*) as count FROM follow WHERE follower_id={session_id} AND following_id='{following_id}'")
+    
+    if result[0]['count'] == 0 :
+        sql.insert(f"INSERT INTO follow(follower_id, following_id) VALUES ('{session_id}', '{following_id}')")
+        return {"result" : "follow"}
+    else : 
+        sql.delete(f"DELETE FROM follow WHERE follower_id='{session_id}' AND following_id='{following_id}'")
+        return {"result" : "unfollow"}
